Extract visibility component
